+++ test/integration/foo-packages/no-config/index.js
@@ -1,0 +1,10 @@
+
+'use strict'
+
+const sum = function (x, y) {
+  return x + y
+}
+
+module.exports = {
+  sum: sum
+}