+++ test/unit/lib/local.specs.js
@@ -1,6 +1,5 @@
 
 const Boom = require('boom')
-const _ = require('lodash')
 const mochaSinonChaiRunner = require('mocha-sinon-chai/runner')
 
 const test = require('../../../index')
@@ -9,8 +8,13 @@ const fixtures = require('../fixtures')
 
 const local = require('../../../lib/local')
 
+const commands = require('../../../lib/commands')
 const options = require('../../../lib/options')
 const treeKill = require('../../../lib/tree-kill')
+
+const deepClone = function (obj) {
+  return JSON.parse(JSON.stringify(obj))
+}
 
 const StdinOnFake = function (options) {
   options = options || {}
@@ -36,8 +40,13 @@ test.describe('local', () => {
     let waitOnMock
     let childProcessMock
     let stdinOnFake
+    let localSuiteFixture
+    let localSuiteWithNoServiceFixture
 
     test.beforeEach(() => {
+      localSuiteFixture = deepClone(fixtures.config.localSuite)
+      localSuiteWithNoServiceFixture = deepClone(fixtures.config.localSuiteWithNoService)
+
       sandbox = test.sinon.sandbox.create()
       tracerMock = new mocks.Tracer()
       waitOnMock = new mocks.WaitOn()
@@ -48,6 +57,9 @@ test.describe('local', () => {
         local: 'fooService'
       })
       sandbox.stub(treeKill, 'kill')
+      sandbox.stub(commands, 'run').usingPromise().resolves({
+        on: childProcessMock.stubs.spawn.on.fake
+      })
       sandbox.stub(process.stdin, 'setRawMode')
       sandbox.stub(process.stdin, 'resume')
       sandbox.stub(process.stdin, 'setEncoding')
@@ -66,35 +78,78 @@ test.describe('local', () => {
     })
 
     test.it('should return a promise', (done) => {
-      const execution = local.run(fixtures.config.localSuite)
+      const execution = local.run(localSuiteFixture)
         .then(() => {
           test.expect(execution).to.be.an.instanceof(Promise)
           done()
         })
     })
 
-    test.it('should execute "before" command as a childProcess synchronous in shell if it is defined in suite', () => {
+    test.describe('when runs "before" command', () => {
       const commandPath = 'fooBeforeLocalCommand'
-      options.get.resolves({
+      const cleanEnv = {
+        fooEnv1: 'fooEnv1',
+        fooEnv2: 'fooEnv2'
+      }
+
+      test.beforeEach(() => {
+        options.get.resolves({
+        })
+        localSuiteFixture.before = {
+          local: {
+            command: commandPath,
+            env: cleanEnv
+          }
+        }
       })
-      return local.run(_.extend({}, fixtures.config.localSuite, {
-        before: {
+
+      test.it('should not execute the "before" command if it is running an specific service or test', () => {
+        options.get.resolves({
+          local: 'fooService'
+        })
+        return local.run(localSuiteFixture).then(() => {
+          return Promise.all([
+            test.expect(commands.run).to.have.been.calledOnce(),
+            test.expect(commands.run).to.not.have.been.calledWith(commandPath)
+          ])
+        })
+      })
+
+      test.it('should execute the "before" command if it is defined in suite', () => {
+        localSuiteFixture.before = {
           local: {
             command: commandPath
           }
         }
-      })).then(() => {
-        return Promise.all([
-          test.expect(childProcessMock.stubs.execFileSync).to.have.been.calledWith(commandPath),
-          test.expect(childProcessMock.stubs.execFileSync.getCall(0).args[2].cwd).to.equal(process.cwd()),
-          test.expect(childProcessMock.stubs.execFileSync.getCall(0).args[2].shell).to.be.true()
-        ])
+        return local.run(localSuiteFixture).then(() => {
+          return Promise.all([
+            test.expect(commands.run).to.have.been.calledWith(commandPath),
+            test.expect(commands.run.getCall(0).args[1].sync).to.be.true()
+          ])
+        })
       })
-    })
 
-    test.it('should not execute the "before" command if it is running an specific service or test', () => {
-      return local.run(fixtures.config.localSuite).then(() => {
-        return test.expect(childProcessMock.stubs.execFileSync).to.not.have.been.called()
+      test.it('should add the suite details to environment variables', () => {
+        const fooSuiteTypeName = 'fooTypeName'
+        return local.run(localSuiteFixture, fooSuiteTypeName).then(() => {
+          const envValues = commands.run.getCall(0).args[1].env
+          return Promise.all([
+            test.expect(envValues.narval_is_docker).to.equal(false),
+            test.expect(envValues.narval_service).to.equal('clean'),
+            test.expect(envValues.narval_suite).to.equal('fooLocalSuite'),
+            test.expect(envValues.narval_suite_type).to.equal(fooSuiteTypeName)
+          ])
+        })
+      })
+
+      test.it('should add custom environment variables', () => {
+        return local.run(localSuiteFixture).then(() => {
+          const envValues = commands.run.getCall(0).args[1].env
+          return Promise.all([
+            test.expect(envValues.fooEnv1).to.equal(cleanEnv.fooEnv1),
+            test.expect(envValues.fooEnv2).to.equal(cleanEnv.fooEnv2)
+          ])
+        })
       })
     })
 
@@ -102,37 +157,34 @@ test.describe('local', () => {
       options.get.resolves({
         local: 'test'
       })
-      return local.run(fixtures.config.localSuite).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return Promise.all([
           test.expect(mochaSinonChaiRunner.run).to.have.been.called(),
-          test.expect(childProcessMock.stubs.execFile).to.not.have.been.called()
+          test.expect(commands.run).to.not.have.been.called()
         ])
       })
     })
 
     test.it('should execute only an specific service if it is defined in "local" option', () => {
-      return local.run(fixtures.config.localSuite).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return Promise.all([
           test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
-          test.expect(childProcessMock.stubs.execFile).to.have.been.calledOnce()
+          test.expect(commands.run).to.have.been.calledOnce()
         ])
       })
     })
 
     test.it('should execute an specific service with coverage if "coverage.from" option is defined in configuration', () => {
       const fakeServiceName = 'fooService2'
-      const suiteFixture = _.extend({}, fixtures.config.localSuite, {
-        coverage: {
-          from: fakeServiceName
-        }
-      })
+      localSuiteFixture.coverage = {
+        from: fakeServiceName
+      }
       options.get.resolves({
         local: fakeServiceName
       })
-      return local.run(suiteFixture).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return Promise.all([
-          test.expect(childProcessMock.stubs.execFile).to.not.have.been.called(),
-          test.expect(childProcessMock.stubs.execFileSync).to.not.have.been.called(),
+          test.expect(commands.run).to.not.have.been.called(),
           test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
           test.expect(childProcessMock.stubs.fork).to.have.been.called()
         ])
@@ -141,19 +193,16 @@ test.describe('local', () => {
 
     test.it('should execute an specific service without coverage if "coverage.enabled" option is false even when coverage.from is defined in configuration', () => {
       const fakeServiceName = 'fooService2'
-      const suiteFixture = _.extend({}, fixtures.config.localSuite, {
-        coverage: {
-          from: fakeServiceName,
-          enabled: false
-        }
-      })
+      localSuiteFixture.coverage = {
+        from: fakeServiceName,
+        enabled: false
+      }
       options.get.resolves({
         local: fakeServiceName
       })
-      return local.run(suiteFixture).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return Promise.all([
-          test.expect(childProcessMock.stubs.execFile).to.have.been.called(),
-          test.expect(childProcessMock.stubs.execFileSync).to.not.have.been.called(),
+          test.expect(commands.run).to.have.been.calledOnce(),
           test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
           test.expect(childProcessMock.stubs.fork).to.not.have.been.called()
         ])
@@ -161,10 +210,10 @@ test.describe('local', () => {
     })
 
     test.it('should execute an specific service without coverage by default', () => {
-      return local.run(fixtures.config.localSuite).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return Promise.all([
-          test.expect(childProcessMock.stubs.execFile.getCall(0).args[0]).to.equal('foo-local-command'),
-          test.expect(childProcessMock.stubs.execFileSync).to.not.have.been.called(),
+          test.expect(commands.run).to.have.been.calledOnce(),
+          test.expect(commands.run.getCall(0).args[0]).to.equal('foo-local-command'),
           test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
           test.expect(childProcessMock.stubs.fork).to.not.have.been.called()
         ])
@@ -176,15 +225,14 @@ test.describe('local', () => {
       options.get.resolves({
         local: fakeServiceName
       })
-      return local.run(fixtures.config.localSuite)
+      return local.run(localSuiteFixture)
         .then(() => {
           return Promise.reject(new Error())
         }).catch((error) => {
           return Promise.all([
             test.expect(Boom.isBoom(error)).to.be.true(),
             test.expect(tracerMock.stubs.error.getCall(0).args[0]).to.contain(fakeServiceName),
-            test.expect(childProcessMock.stubs.execFile).to.not.have.been.called(),
-            test.expect(childProcessMock.stubs.execFileSync).to.not.have.been.called(),
+            test.expect(commands.run).to.not.have.been.called(),
             test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
             test.expect(childProcessMock.stubs.fork).to.not.have.been.called()
           ])
@@ -193,24 +241,22 @@ test.describe('local', () => {
 
     test.it('should reject the promise with a controlled error if a service has not defined a command to execute in local', () => {
       const fakeServiceName = 'fooService2'
-      const suiteFixture = _.extend({}, fixtures.config.localSuite, {
-        services: [
-          {
-            name: fakeServiceName
-          }
-        ]
-      })
+      localSuiteFixture.services = [
+        {
+          name: fakeServiceName
+        }
+      ]
       options.get.resolves({
         local: fakeServiceName
       })
-      return local.run(suiteFixture)
+      return local.run(localSuiteFixture)
         .then(() => {
           return Promise.reject(new Error())
         })
         .catch((err) => {
           return Promise.all([
             test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
-            test.expect(childProcessMock.stubs.execFile).to.not.have.been.called(),
+            test.expect(commands.run).to.not.have.been.called(),
             test.expect(tracerMock.stubs.error.getCall(0).args[0]).to.contain(fakeServiceName),
             test.expect(Boom.isBoom(err)).to.be.true()
           ])
@@ -218,51 +264,53 @@ test.describe('local', () => {
     })
 
     test.it('should print a debug trace when an specific service is defined in "local" option and it has finished', () => {
-      return local.run(fixtures.config.localSuite).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return test.expect(tracerMock.stubs.debug.getCall(tracerMock.stubs.debug.callCount - 1).args[0]).to.contain(`Service "fooService" closed`)
       })
     })
 
     test.it('should reject the promise if the service execution fails, adding the service name to the error message', () => {
-      childProcessMock.stubs.execFile.on.returns(1)
-      return local.run(fixtures.config.localSuite).catch((err) => {
-        return test.expect(err.message).to.contain('fooService')
-      })
+      childProcessMock.stubs.spawn.on.returns(1)
+      return local.run(localSuiteFixture)
+        .then(() => {
+          return test.expect(false).to.be.true()
+        })
+        .catch((err) => {
+          return test.expect(err.message).to.contain('fooService')
+        })
     })
 
     test.it('should run all services and test when no specific test or service is defined in "local" option', () => {
       options.get.resolves({})
-      return local.run(fixtures.config.localSuite).then(() => {
+      return local.run(localSuiteFixture).then(() => {
         return Promise.all([
           test.expect(mochaSinonChaiRunner.run).to.have.been.called(),
-          test.expect(childProcessMock.stubs.execFile).to.have.been.calledTwice()
+          test.expect(commands.run).to.have.been.calledTwice()
         ])
       })
     })
 
     test.describe('when runs all services and test', () => {
       const coveragedService = 'fooService2'
-      const suiteFixture = _.extend({}, fixtures.config.localSuite, {
-        coverage: {
+      test.beforeEach(() => {
+        localSuiteFixture.coverage = {
           from: coveragedService
-        },
-        services: JSON.parse(JSON.stringify(fixtures.config.localSuite.services)).concat([{
+        }
+        localSuiteFixture.services = JSON.parse(JSON.stringify(localSuiteFixture.services)).concat([{
           name: 'fooService3',
           local: {
             command: 'foo-local-command-3'
           }
         }])
-      })
-      test.beforeEach(() => {
         options.get.resolves({})
       })
 
       test.it('should first start all services, then run test', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return Promise.all([
-            test.expect(childProcessMock.stubs.execFile).to.have.been.calledTwice(),
+            test.expect(commands.run).to.have.been.calledTwice(),
             test.expect(childProcessMock.stubs.fork).to.have.been.calledTwice(),
-            test.expect(childProcessMock.stubs.execFile.getCall(0).args[0]).to.contain('foo-local-command'),
+            test.expect(commands.run.getCall(0).args[0]).to.contain('foo-local-command'),
             test.expect(childProcessMock.stubs.fork.getCall(0).args[0]).to.contain('msc-istanbul.js'),
             test.expect(childProcessMock.stubs.fork.getCall(0).args[2].env.servicePath).to.contain('foo-local-command2.js'),
             test.expect(childProcessMock.stubs.fork.getCall(1).args[0]).to.contain('msc_mocha.js')
@@ -271,10 +319,10 @@ test.describe('local', () => {
       })
 
       test.it('should run only test when suite has no services', () => {
-        return local.run(fixtures.config.localSuiteWithNoService).then(() => {
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return Promise.all([
             test.expect(childProcessMock.stubs.fork).to.not.have.been.called(),
-            test.expect(childProcessMock.stubs.execFile).to.not.have.been.called(),
+            test.expect(commands.run).to.not.have.been.called(),
             test.expect(mochaSinonChaiRunner.run).to.have.been.called()
           ])
         })
@@ -282,7 +330,7 @@ test.describe('local', () => {
 
       test.it('should reject the promise if the test execution fails', () => {
         mochaSinonChaiRunner.run.rejects(new Error())
-        return local.run(fixtures.config.localSuiteWithNoService).then(() => {
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return test.expect(false).to.be.true()
         }).catch((error) => {
           return Promise.all([
@@ -293,13 +341,13 @@ test.describe('local', () => {
       })
 
       test.it('should kill all not coveraged services when test finish', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return test.expect(treeKill.kill).to.have.been.calledTwice()
         })
       })
 
       test.it('should send an exit signal to coveraged services when test finish', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return Promise.all([
             test.expect(childProcessMock.stubs.fork.send).to.have.been.calledOnce(),
             test.expect(childProcessMock.stubs.fork.send).to.have.been.calledWith({exit: true})
@@ -308,7 +356,7 @@ test.describe('local', () => {
       })
 
       test.it('should print a debug trace for each closed service', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return test.expect(tracerMock.stubs.debug.callCount).to.equal(8)
         })
       })
@@ -316,7 +364,7 @@ test.describe('local', () => {
 
     test.it('should run coveraged tests when coverage is not specified for a service in configuration', () => {
       options.get.resolves({})
-      return local.run(fixtures.config.localSuiteWithNoService).then(() => {
+      return local.run(localSuiteWithNoServiceFixture).then(() => {
         return Promise.all([
           test.expect(mochaSinonChaiRunner.run).to.have.been.called(),
           test.expect(childProcessMock.stubs.fork).to.not.have.been.called()
@@ -325,13 +373,11 @@ test.describe('local', () => {
     })
 
     test.it('should run coveraged tests when coverage is specified for test in configuration', () => {
-      const suite = _.extend({}, fixtures.config.localSuiteWithNoService, {
-        coverage: {
-          from: 'test'
-        }
-      })
+      localSuiteWithNoServiceFixture.coverage = {
+        from: 'test'
+      }
       options.get.resolves({})
-      return local.run(suite).then(() => {
+      return local.run(localSuiteWithNoServiceFixture).then(() => {
         return Promise.all([
           test.expect(mochaSinonChaiRunner.run).to.have.been.called(),
           test.expect(childProcessMock.stubs.fork).to.not.have.been.called()
@@ -340,14 +386,12 @@ test.describe('local', () => {
     })
 
     test.it('should run not coveraged tests when coverage is disabled in configuration', () => {
-      const suite = _.extend({}, fixtures.config.localSuiteWithNoService, {
-        coverage: {
-          enabled: false,
-          from: 'test'
-        }
-      })
+      localSuiteWithNoServiceFixture.coverage = {
+        enabled: false,
+        from: 'test'
+      }
       options.get.resolves({})
-      return local.run(suite).then(() => {
+      return local.run(localSuiteWithNoServiceFixture).then(() => {
         return Promise.all([
           test.expect(mochaSinonChaiRunner.run).to.not.have.been.called(),
           test.expect(childProcessMock.stubs.fork).to.have.been.called()
@@ -358,7 +402,7 @@ test.describe('local', () => {
     test.it('should reject the promise if the test execution fails, specifying it in the error message', () => {
       options.get.resolves({})
       mochaSinonChaiRunner.run.rejects(new Error())
-      return local.run(fixtures.config.localSuiteWithNoService)
+      return local.run(localSuiteWithNoServiceFixture)
         .then(() => {
           return test.expect(false).to.be.true()
         })
@@ -371,32 +415,55 @@ test.describe('local', () => {
     })
 
     test.describe('when runs coveraged tests', () => {
+      const fooSuiteType = 'fooType'
+
       test.beforeEach(() => {
         options.get.resolves({})
       })
 
+      test.it('should add the suite details to environment variables', () => {
+        return local.run(localSuiteWithNoServiceFixture, fooSuiteType).then(() => {
+          let envValues = mochaSinonChaiRunner.run.getCall(0).args[1].env
+          return Promise.all([
+            test.expect(envValues.narval_is_docker).to.equal(false),
+            test.expect(envValues.narval_service).to.equal('test'),
+            test.expect(envValues.narval_suite).to.equal('fooLocalSuite2'),
+            test.expect(envValues.narval_suite_type).to.equal(fooSuiteType)
+          ])
+        })
+      })
+
+      test.it('should add custom environment variables', () => {
+        localSuiteWithNoServiceFixture.test.local = {
+          env: {
+            customVar: 'custom value'
+          }
+        }
+        return local.run(localSuiteWithNoServiceFixture, fooSuiteType).then(() => {
+          return test.expect(mochaSinonChaiRunner.run.getCall(0).args[1].env.customVar).to.equal('custom value')
+        })
+      })
+
       test.it('should print a debug message with details about execution type', () => {
-        return local.run(fixtures.config.localSuiteWithNoService).then(() => {
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return test.expect(tracerMock.stubs.debug.getCall(0).args[0]).to.contain('coverage enabled')
         })
       })
 
       test.it('should call to mocha-sinon-chai runner, passing the istanbul and mocha configuration', () => {
-        const suite = _.extend({}, fixtures.config.localSuiteWithNoService, {
-          test: {
-            specs: 'foo/path/specs',
-            config: {
-              fooMochaParam1: 'foo',
-              fooParam2: 'fake'
-            }
-          },
-          coverage: {
-            config: {
-              fooIstanbulParam: 'fooValue'
-            }
+        localSuiteWithNoServiceFixture.test = {
+          specs: 'foo/path/specs',
+          config: {
+            fooMochaParam1: 'foo',
+            fooParam2: 'fake'
           }
-        })
-        return local.run(suite).then(() => {
+        }
+        localSuiteWithNoServiceFixture.coverage = {
+          config: {
+            fooIstanbulParam: 'fooValue'
+          }
+        }
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return Promise.all([
             test.expect(mochaSinonChaiRunner.run).to.have.been.called(),
             test.expect(mochaSinonChaiRunner.run.getCall(0).args[0]).to.equal('--istanbul --include-all-sources --root=. --colors --print=summary --dir=.coverage/fooLocalSuite2/fooLocalSuite2 --fooIstanbulParam=fooValue --mocha --recursive --colors --reporter spec --fooMochaParam1 foo --fooParam2 fake foo/path/specs')
@@ -406,31 +473,57 @@ test.describe('local', () => {
     })
 
     test.describe('when runs not coveraged tests', () => {
-      const suite = _.extend({}, fixtures.config.localSuiteWithNoService, {
-        test: {
+      const fooSuiteType = 'foo-suite-type'
+      test.beforeEach(() => {
+        localSuiteWithNoServiceFixture.test = {
           specs: 'foo2/specs',
           config: {
             fooMochaParam1: 'foo',
             fooParam2: 'fake'
           }
-        },
-        coverage: {
+        }
+        localSuiteWithNoServiceFixture.coverage = {
           enabled: false
         }
-      })
-
-      test.beforeEach(() => {
         options.get.resolves({})
       })
 
+      test.it('should add the suite details to environment variables', () => {
+        return local.run(localSuiteWithNoServiceFixture, fooSuiteType).then(() => {
+          let envValues = childProcessMock.stubs.fork.getCall(0).args[2].env
+          return Promise.all([
+            test.expect(envValues.narval_is_docker).to.equal(false),
+            test.expect(envValues.narval_service).to.equal('test'),
+            test.expect(envValues.narval_suite).to.equal('fooLocalSuite2'),
+            test.expect(envValues.narval_suite_type).to.equal(fooSuiteType)
+          ])
+        })
+      })
+
+      test.it('should add custom environment variables', () => {
+        localSuiteWithNoServiceFixture.test.local = {
+          env: {
+            var1: 'value 1',
+            var2: 'value 2'
+          }
+        }
+        return local.run(localSuiteWithNoServiceFixture, fooSuiteType).then(() => {
+          const envValues = childProcessMock.stubs.fork.getCall(0).args[2].env
+          return Promise.all([
+            test.expect(envValues.var1).to.equal('value 1'),
+            test.expect(envValues.var2).to.equal('value 2')
+          ])
+        })
+      })
+
       test.it('should print a debug message with details about execution type', () => {
-        return local.run(suite).then(() => {
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return test.expect(tracerMock.stubs.debug.getCall(0).args[0]).to.contain('without coverage')
         })
       })
 
       test.it('should open a mocha child process fork, passing the mocha configuration', () => {
-        return local.run(suite).then(() => {
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return Promise.all([
             test.expect(childProcessMock.stubs.fork).to.have.been.calledOnce(),
             test.expect(childProcessMock.stubs.fork.getCall(0).args[0]).to.contain('msc_mocha.js'),
@@ -450,16 +543,16 @@ test.describe('local', () => {
       })
 
       test.it('should resolve the promise when mocha execution finish OK', () => {
-        return local.run(suite).then(() => {
+        return local.run(localSuiteWithNoServiceFixture).then(() => {
           return test.expect(true).to.be.true()
         })
       })
 
       test.it('should reject the promise when mocha execution fails', () => {
         childProcessMock.stubs.fork.on.returns(1)
-        return local.run(suite)
+        return local.run(localSuiteWithNoServiceFixture)
           .then(() => {
-            throw new Error()
+            return test.expect(false).to.be.true()
           })
           .catch((error) => {
             return Promise.all([
@@ -471,7 +564,7 @@ test.describe('local', () => {
     })
 
     test.it('should not call to wait-on when test has not a wait-for property in configuration', () => {
-      return local.run(fixtures.config.localSuite)
+      return local.run(localSuiteFixture)
         .then(() => {
           return test.expect(waitOnMock.stubs.wait).to.not.have.been.called()
         })
@@ -479,21 +572,19 @@ test.describe('local', () => {
 
     test.describe('when test has a wait-for property in configuration', () => {
       const fooServiceUrl = 'http://fake-service:3000'
-      const suite = _.extend({}, fixtures.config.localSuite, {
-        test: {
+
+      test.beforeEach(() => {
+        localSuiteFixture.test = {
           specs: 'foo/wait/specs',
           local: {
             'wait-for': fooServiceUrl
           }
         }
-      })
-
-      test.beforeEach(() => {
         options.get.resolves({})
       })
 
       test.it('should print a debug trace', () => {
-        return local.run(suite)
+        return local.run(localSuiteFixture)
           .then(() => {
             return Promise.all([
               test.expect(tracerMock.stubs.debug.getCall(2).args[0]).to.contain('Waiting'),
@@ -503,7 +594,7 @@ test.describe('local', () => {
       })
 
       test.it('should call to waitOn, passing the configuration', () => {
-        return local.run(suite)
+        return local.run(localSuiteFixture)
           .then(() => {
             return Promise.all([
               test.expect(waitOnMock.stubs.wait).to.have.been.called(),
@@ -515,7 +606,7 @@ test.describe('local', () => {
       test.it('should reject the promise if it receives an error from waitOn execution', () => {
         const fooErrorMessage = 'Wait on error'
         waitOnMock.stubs.wait.returns(new Error(fooErrorMessage))
-        return local.run(suite)
+        return local.run(localSuiteFixture)
           .then(() => {
             throw new Error()
           })
@@ -529,60 +620,52 @@ test.describe('local', () => {
     })
 
     test.describe('when runs a not coveraged service', () => {
-      test.it('should log the data received from the execution, aplying a trim function', () => {
-        const fooData = 'foo process data'
-        childProcessMock.stubs.execFile.stdout.on.returns(`   ${fooData}    `)
-        return local.run(fixtures.config.localSuite).then(() => {
-          return test.expect(console.log).to.have.been.calledWith(fooData)
+      const fooSuiteTypeName = 'fooTypeName'
+
+      test.it('should add the suite details to environment variables', () => {
+        return local.run(localSuiteFixture, fooSuiteTypeName).then(() => {
+          const envValues = commands.run.getCall(0).args[1].env
+          return Promise.all([
+            test.expect(commands.run).to.have.been.calledOnce(),
+            test.expect(commands.run.getCall(0).args[0]).to.equal('foo-local-command'),
+            test.expect(envValues.narval_is_docker).to.equal(false),
+            test.expect(envValues.narval_service).to.equal('fooService'),
+            test.expect(envValues.narval_suite).to.equal('fooLocalSuite'),
+            test.expect(envValues.narval_suite_type).to.equal(fooSuiteTypeName)
+          ])
         })
       })
 
-      test.it('should log the errors received from the execution, aplying a trim function', () => {
-        const fooData = 'foo error data'
-        childProcessMock.stubs.execFile.stderr.on.runOnRegister(true)
-        childProcessMock.stubs.execFile.stderr.on.returns(`   ${fooData}    `)
-        return local.run(fixtures.config.localSuite).then(() => {
-          return test.expect(console.log).to.have.been.calledWith(fooData)
-        })
-      })
-
-      test.it('should not log empty data received from the execution', () => {
-        const fooData = '   '
-        childProcessMock.stubs.execFile.stdout.on.returns(fooData)
-        return local.run(fixtures.config.localSuite).then(() => {
-          return test.expect(console.log).to.not.have.been.calledWith(fooData)
-        })
-      })
-
-      test.it('should not log empty data received from the execution', () => {
-        const fooData = '   '
-        childProcessMock.stubs.execFile.stdout.on.returns(fooData)
-        return local.run(fixtures.config.localSuite).then(() => {
-          return test.expect(console.log).to.not.have.been.calledWith(fooData)
+      test.it('should add custom environment variables', () => {
+        return local.run(localSuiteFixture, fooSuiteTypeName).then(() => {
+          const envValues = commands.run.getCall(0).args[1].env
+          return Promise.all([
+            test.expect(envValues.fooEnvVar1).to.equal('fooEnvironment var 1'),
+            test.expect(envValues.fooEnvVar2).to.equal('fooEnv2')
+          ])
         })
       })
     })
 
     test.describe('when runs a coveraged service', () => {
       const fakeServiceName = 'fooService2'
-      const suiteFixture = _.extend({}, fixtures.config.localSuite, {
-        coverage: {
+      const fooSuiteTypeName = 'FooType'
+
+      test.beforeEach(() => {
+        localSuiteFixture.coverage = {
           from: fakeServiceName,
           config: {
             fooConfig1: 'config1',
             fooConfig2: 'config2'
           }
         }
-      })
-
-      test.beforeEach(() => {
         options.get.resolves({
           local: fakeServiceName
         })
       })
 
-      test.it('should set the process stdin to raw mode, in order to intercept CTRL-C and stop only the coveraged service', () => {
-        return local.run(suiteFixture).then(() => {
+      test.it('should set the process stdin to raw mode, in order to intercept CTRL-C and stop only the coveraged service when service is started alone', () => {
+        return local.run(localSuiteFixture).then(() => {
           return Promise.all([
             test.expect(process.stdin.setRawMode).to.have.been.called(),
             test.expect(process.stdin.resume).to.have.been.called()
@@ -590,8 +673,20 @@ test.describe('local', () => {
         })
       })
 
+      test.it('should not set the process stdin to raw mode when all suite is ran', () => {
+        options.get.resolves({
+          local: true
+        })
+        return local.run(localSuiteFixture).then(() => {
+          return Promise.all([
+            test.expect(process.stdin.setRawMode).to.not.have.been.called(),
+            test.expect(process.stdin.resume).to.not.have.been.called()
+          ])
+        })
+      })
+
       test.it('should fork an istanbul child process, passing to it the istanbul and command arguments', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           let forkCall = childProcessMock.stubs.fork.getCall(0)
           return Promise.all([
             test.expect(forkCall.args[0]).to.contain('msc-istanbul.js'),
@@ -604,14 +699,32 @@ test.describe('local', () => {
       })
 
       test.it('should set the command path to be executed as "servicePath" environment variable, and pass it to the fork execution', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return test.expect(childProcessMock.stubs.fork.getCall(0).args[2].env.servicePath).to.contain('foo-local-command2.js')
+        })
+      })
+
+      test.it('should add the suite details to environment variables', () => {
+        return local.run(localSuiteFixture, fooSuiteTypeName).then(() => {
+          const envValues = childProcessMock.stubs.fork.getCall(0).args[2].env
+          return Promise.all([
+            test.expect(envValues.narval_is_docker).to.equal(false),
+            test.expect(envValues.narval_service).to.equal('fooService2'),
+            test.expect(envValues.narval_suite).to.equal('fooLocalSuite'),
+            test.expect(envValues.narval_suite_type).to.equal(fooSuiteTypeName)
+          ])
+        })
+      })
+
+      test.it('should add custom environment variables', () => {
+        return local.run(localSuiteFixture, fooSuiteTypeName).then(() => {
+          return test.expect(childProcessMock.stubs.fork.getCall(0).args[2].env.fooEnv).to.equal('fooEnv value')
         })
       })
 
       test.it('should intercept the CTRL-C and send and exit signal to service', () => {
         stdinOnFake.returns('\u0003')
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return Promise.all([
             test.expect(childProcessMock.stubs.fork.send).to.have.been.calledWith({exit: true}),
             test.expect(tracerMock.stubs.debug.getCall(1).args[0]).to.contain('CTRL-C')
@@ -620,7 +733,7 @@ test.describe('local', () => {
       })
 
       test.it('should restore the stdin raw mode, and stop intercepting CTRL-C when process finish', () => {
-        return local.run(suiteFixture).then(() => {
+        return local.run(localSuiteFixture).then(() => {
           return Promise.all([
             test.expect(process.stdin.setRawMode).to.have.been.calledTwice(),
             test.expect(process.stdin.pause).to.have.been.called()